Tokens added for all controllers

+++ SoftwareLocation/Web/Controllers/DepartmentsController.cs
@@ -14,39 +14,79 @@ namespace Web.Controllers
     public class DepartmentsController : ControllerBase
     {
         IDepartmentService _departmentService;
+        IAuthorityService _authorityService;
 
-        public DepartmentsController(IDepartmentService departmentService)
+        public DepartmentsController(IDepartmentService departmentService, IAuthorityService authorityService)
         {
             _departmentService = departmentService;
+            _authorityService = authorityService;
         }
 
         [HttpGet("getall")]
-        public List<Department> GetAll()
+        public List<Department> GetAll(string token)
         {
-          var result = _departmentService.GetAll();
-            return result;
+            if (_authorityService.GetByToken(token)!=null)
+            {
+                var result = _departmentService.GetAll();
+                return result;
+            }
+            else
+            {
+                return null;
+            }
         }
 
         [HttpGet("getbyid")]
-        public Department GetById(int id)
+        public Department GetById(int id, string token)
         {
-            var result = _departmentService.GetById(id);
-            return result;
+            if (_authorityService.GetByToken(token) != null)
+            {
+                var result = _departmentService.GetById(id);
+                return result;
+            }
+            else
+            {
+                return null;
+            }
         }
 
         [HttpPost("add")]
-        public void Add(Department department) {
-            _departmentService.Add(department);
+        public string Add(Department department , string token) {
+            if (_authorityService.GetByToken(token) != null)
+            {
+                 _departmentService.Add(department);
+                return "Department Added Successfully";
+            }
+            else
+            {
+                return "Access Denied";
+            }
         }
 
         [HttpPost("update")]
-        public void Update(Department department) {
-            _departmentService.Update(department);
+        public string Update(Department department , string token) {
+            if (_authorityService.GetByToken(token) != null)
+            {
+                _departmentService.Update(department);
+                return "Department Updated Successfully";
+            }
+            else
+            {
+                return "Access Denied";
+            }
         }
 
         [HttpDelete("delete")]
-        public void Delete(Department department) {
-            _departmentService.Delete(department);
+        public string Delete(Department department , string token) {
+            if (_authorityService.GetByToken(token) != null)
+            {
+                _departmentService.Delete(department);
+                return "Department Deleted Successfully";
+            }
+            else
+            {
+                return "Access Denied";
+            }
         }
 
     }
